+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 * Added new constructor overloads for `AtProtoHttpClient` to allow for use with `MetricsFactory`.
 * Added `Throttled` to `AccountStatus` in Jetstream account events.
 * Added `MaxMessageSize` to `JetStreamOptions` to guard against a malicious jetstream server sending overly large messages.
-* Added optional validation callbacks to `BuildOAuthLoginUri` to allow for validation of the discovered PDS and authorization server URIs.
+* Added optional validation callbacks to `AtProtoAgent.BuildOAuthLoginUri` to allow for validation of the discovered PDS and authorization server URIs.
+* Added default validation function to disallow loopback and private IP ranges in discovered URIs, to mitigate SSRF attacks.
 * Added override on `ToString()` on `AtProtoCredential` to return a redacted string in case of accidental logging.
 
 ### idunno.AtProto.Types